+++ src/Image/Infrastructure/VectorStorage/LibraryImageUpdater.php
@@ -11,6 +11,7 @@ use ChronicleKeeper\Image\Domain\Entity\Image;
 use ChronicleKeeper\Library\Infrastructure\VectorStorage\VectorImage;
 use ChronicleKeeper\Shared\Application\Query\QueryService;
 use ChronicleKeeper\Shared\Infrastructure\LLMChain\EmbeddingCalculator;
+use ChronicleKeeper\Shared\Infrastructure\LLMChain\Exception\EmbeddingCalculationFailed;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
 
@@ -63,8 +64,24 @@ class LibraryImageUpdater
     {
         // Calculate Content Chunks
         $contentChunks = $this->embeddingCalculator->createTextChunks($image->getDescription());
-        // Calculate vectors from the chunks
-        $vectors = $this->embeddingCalculator->getMultipleEmbeddings($contentChunks);
+
+        if (count($contentChunks) === 0) {
+            return [];
+        }
+
+        try {
+            $vectors = $this->embeddingCalculator->getMultipleEmbeddings($contentChunks);
+        } catch (EmbeddingCalculationFailed $e) {
+            $this->logger->error(
+                'The creation of embedding chunks for document "{image}" failed.',
+                [
+                    'message' => $e->getMessage(),
+                    'image' => $image->getId(),
+                ],
+            );
+
+            return [];
+        }
 
         $vectorImages = [];
         foreach ($contentChunks as $index => $chunk) {